+++ cmd/issuer/main.go
@@ -28,7 +28,7 @@ import (
 	vaultfs "github.com/agntcy/identity/internal/issuer/vault/data/filesystem"
 	"github.com/agntcy/identity/internal/issuer/verify"
 	"github.com/agntcy/identity/internal/pkg/nodeapi"
-	"github.com/agntcy/identity/internal/pkg/oidc"
+	"github.com/agntcy/identity/pkg/oidc"
 
 	"github.com/spf13/cobra"
 )
