+++ cmd/node/main.go
@@ -20,11 +20,11 @@ import (
 	"github.com/agntcy/identity/internal/node"
 	nodegrpc "github.com/agntcy/identity/internal/node/grpc"
 	"github.com/agntcy/identity/internal/pkg/grpcutil"
-	"github.com/agntcy/identity/internal/pkg/oidc"
 	"github.com/agntcy/identity/pkg/cmd"
 	"github.com/agntcy/identity/pkg/db"
 	"github.com/agntcy/identity/pkg/grpcserver"
 	"github.com/agntcy/identity/pkg/log"
+	"github.com/agntcy/identity/pkg/oidc"
 	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 	"github.com/rs/cors"
 	"github.com/sirupsen/logrus"
