+++ internal/node/id_generator.go
@@ -12,8 +12,8 @@ import (
 	issuerverification "github.com/agntcy/identity/internal/core/issuer/verification"
 	vctypes "github.com/agntcy/identity/internal/core/vc/types"
 	"github.com/agntcy/identity/internal/pkg/errutil"
-	"github.com/agntcy/identity/internal/pkg/oidc"
 	"github.com/agntcy/identity/pkg/log"
+	"github.com/agntcy/identity/pkg/oidc"
 )
 
 // All IDP schemes supported by the ID generator.
