+++ internal/node/vc_service_test.go
@@ -23,8 +23,8 @@ import (
 	vctesting "github.com/agntcy/identity/internal/core/vc/testing"
 	vctypes "github.com/agntcy/identity/internal/core/vc/types"
 	"github.com/agntcy/identity/internal/node"
-	"github.com/agntcy/identity/internal/pkg/oidc"
-	oidctesting "github.com/agntcy/identity/internal/pkg/oidc/testing"
+	"github.com/agntcy/identity/pkg/oidc"
+	oidctesting "github.com/agntcy/identity/pkg/oidc/testing"
 	"github.com/google/uuid"
 	"github.com/lestrrat-go/jwx/v3/jwa"
 	"github.com/lestrrat-go/jwx/v3/jwk"
