+++ pkg/joseutil/joseutil_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/agntcy/identity/internal/core/id/types"
-	"github.com/agntcy/identity/internal/pkg/joseutil"
+	"github.com/agntcy/identity/pkg/joseutil"
 	"github.com/stretchr/testify/assert"
 )
