+++ pkg/joseutil/sign_verify.go
