+++ pkg/keystore/hashicorp_vault_key_service.go
