+++ pkg/keystore/local_file_key_test.go
@@ -9,8 +9,8 @@ import (
 	"os"
 	"testing"
 
-	"github.com/agntcy/identity/internal/core/keystore"
-	"github.com/agntcy/identity/internal/pkg/joseutil"
+	"github.com/agntcy/identity/pkg/joseutil"
+	"github.com/agntcy/identity/pkg/keystore"
 	"github.com/stretchr/testify/assert"
 )
